+++ include/arion/platforms/linux/lnx_arch_manager.hpp
@@ -1,0 +1,14 @@
+#ifndef ARION_LNX_ARCH_MANAGER_HPP
+#define ARION_LNX_ARCH_MANAGER_HPP
+
+#include <arion/common/global_defs.hpp>
+#include <memory>
+#include <vector>
+
+class LinuxArchManager
+{
+  public:
+    virtual std::unique_ptr<std::map<arion::REG, arion::RVAL>> prstatus_to_regs(std::vector<arion::BYTE> prstatus) = 0;
+};
+
+#endif // ARION_LNX_ARCH_MANAGER_HPP