+++ src/common/code_tracer.cpp
@@ -65,7 +65,7 @@ void CodeTracer::prepare_file()
         if (this->mode == TRACE_MODE::CTXT)
             regs_sec_off = (off_t)this->out_f.tellp() + sizeof(off_t) * 2;
         this->out_f.write((char *)&regs_sec_off, sizeof(off_t));
-        std::vector<REG> ctxt_regs = arion->abi->get_context_regs();
+        std::vector<REG> ctxt_regs = arion->arch->get_context_regs();
         size_t ctxt_regs_sz = ctxt_regs.size();
         off_t data_sec_off = (off_t)this->out_f.tellp() + sizeof(off_t) + sizeof(size_t) + ctxt_regs_sz * sizeof(REG);
         this->out_f.write((char *)&data_sec_off, sizeof(off_t));
@@ -174,7 +174,7 @@ void CodeTracer::process_hit(ADDR addr, size_t sz)
     this->total_hits++;
     std::unique_ptr<CODE_HIT> hit = std::make_unique<CODE_HIT>(addr - mapping_start, sz, mod_id);
     if (this->mode == TRACE_MODE::CTXT)
-        hit->regs = std::move(arion->abi->dump_regs());
+        hit->regs = std::move(arion->arch->dump_regs());
     this->hits.push_back(std::move(hit));
 
     size_t hits_sz = this->hits.size();