+++ src/common/context_manager.cpp
@@ -25,8 +25,8 @@ std::shared_ptr<ARION_CONTEXT> ContextManager::save()
         std::unique_ptr<ARION_THREAD> arion_t_cpy = std::make_unique<ARION_THREAD>(arion_t.second.get());
         if (arion_t_cpy->tid == running_tid)
         {
-            arion_t_cpy->regs_state = arion->abi->dump_regs();
-            arion_t_cpy->tls_addr = arion->abi->dump_tls();
+            arion_t_cpy->regs_state = arion->arch->dump_regs();
+            arion_t_cpy->tls_addr = arion->arch->dump_tls();
         }
         thread_list.push_back(std::move(arion_t_cpy));
     }
@@ -154,8 +154,8 @@ void ContextManager::restore(std::shared_ptr<ARION_CONTEXT> ctx, bool restore_ma
     for (std::unique_ptr<ARION_FUTEX> &arion_f : ctx->futex_list)
         arion->threads->futex_wait(arion_f->tid, arion_f->futex_addr, arion_f->futex_bitmask);
     arion->threads->set_running_tid(ctx->running_tid);
-    arion->abi->load_regs(std::move(arion->threads->threads_map[ctx->running_tid]->regs_state));
-    arion->abi->load_tls(std::move(arion->threads->threads_map[ctx->running_tid]->tls_addr));
+    arion->arch->load_regs(std::move(arion->threads->threads_map[ctx->running_tid]->regs_state));
+    arion->arch->load_tls(std::move(arion->threads->threads_map[ctx->running_tid]->tls_addr));
 }
 
 void ContextManager::restore(std::shared_ptr<ARION_CONTEXT> ctx, std::vector<std::shared_ptr<ARION_MEM_EDIT>> edits)