+++ src/platforms/linux/elf_parser.cpp
@@ -1,10 +1,14 @@
 #include <arion/LIEF/ELF/NoteDetails/core/CoreFile.hpp>
+#include <arion/LIEF/ELF/NoteDetails/core/CorePrStatus.hpp>
 #include <arion/LIEF/ELF/Segment.hpp>
 #include <arion/LIEF/ELF/enums.hpp>
 #include <arion/arion.hpp>
 #include <arion/common/global_defs.hpp>
 #include <arion/common/global_excepts.hpp>
+#include <arion/common/logger.hpp>
 #include <arion/platforms/linux/elf_parser.hpp>
+#include <arion/platforms/linux/lnx_excepts.hpp>
+#include <arion/utils/convert_utils.hpp>
 #include <filesystem>
 #include <memory>
 
@@ -16,6 +20,96 @@ std::map<LIEF::ELF::Header::FILE_TYPE, ELF_FILE_TYPE> lief_arion_file_types = {
     {LIEF::ELF::Header::FILE_TYPE::EXEC, ELF_FILE_TYPE::EXEC},
     {LIEF::ELF::Header::FILE_TYPE::DYN, ELF_FILE_TYPE::DYN},
     {LIEF::ELF::Header::FILE_TYPE::CORE, ELF_FILE_TYPE::CORE}};
+
+void ElfCoredumpParser::parse_file_note(const LIEF::ELF::Note &note,
+                                        std::vector<std::shared_ptr<struct arion::SEGMENT>> segments)
+{
+    if (!LIEF::ELF::CoreFile::classof(&note))
+        return;
+
+    const auto &nt_core_file = static_cast<const LIEF::ELF::CoreFile &>(note);
+    for (auto &file_entry : nt_core_file)
+    {
+        for (std::shared_ptr<struct arion::SEGMENT> &seg : segments)
+        {
+            if (seg->virt_addr >= file_entry.start && seg->virt_addr < file_entry.end)
+                seg->info = file_entry.path;
+        }
+    }
+}
+
+void ElfCoredumpParser::parse_prstatus_note(const LIEF::ELF::Note &note,
+                                            std::shared_ptr<ARION_ELF_PARSER_ATTRIBUTES> attrs)
+{
+    if (!LIEF::ELF::CorePrStatus::classof(&note))
+        return;
+
+    const auto &nt_core_prstatus = static_cast<const LIEF::ELF::CorePrStatus &>(note);
+}
+
+void ElfCoredumpParser::parse_prpsinfo_note(const LIEF::ELF::Note &note,
+                                            std::shared_ptr<ARION_ELF_PARSER_ATTRIBUTES> attrs)
+{
+    std::shared_ptr<Arion> arion = this->arion.lock();
+    if (!arion)
+        throw ExpiredWeakPtrException("Arion");
+
+    if (!LIEF::ELF::CorePrPsInfo::classof(&note))
+        return;
+    if (this->found_prpsinfo)
+        throw MultiplePrPsInfoNotesException();
+    this->found_prpsinfo = true;
+    const auto &nt_core_prpsinfo = static_cast<const LIEF::ELF::CorePrPsInfo &>(note);
+    const auto info = nt_core_prpsinfo.info();
+    std::string program_name = info->args_stripped();
+    if (program_name.length())
+    {
+        attrs->args = std::vector<std::string>({info->args_stripped()});
+        attrs->path = program_name;
+    }
+    arion->set_pid(info->pid);
+    arion->set_pgid(info->pgrp);
+    arion->set_sid(info->sid);
+    if (info->zombie)
+        arion->set_zombie();
+}
+
+void ElfCoredumpParser::parse_fpregset_note(const LIEF::ELF::Note &note)
+{
+}
+
+std::unique_ptr<LIEF::ELF::Binary> ElfCoredumpParser::parse_coredump_data(
+    std::unique_ptr<LIEF::ELF::Binary> elf, std::shared_ptr<ARION_ELF_PARSER_ATTRIBUTES> attrs,
+    std::vector<std::shared_ptr<struct arion::SEGMENT>> segments)
+{
+    std::shared_ptr<Arion> arion = this->arion.lock();
+    if (!arion)
+        throw ExpiredWeakPtrException("Arion");
+
+    for (const LIEF::ELF::Note &note : elf->notes())
+    {
+        switch (note.type())
+        {
+        case LIEF::ELF::Note::TYPE::CORE_FILE:
+            this->parse_file_note(note, segments);
+            break;
+        case LIEF::ELF::Note::TYPE::CORE_PRSTATUS:
+            this->parse_prstatus_note(note, attrs);
+            break;
+        case LIEF::ELF::Note::TYPE::CORE_PRPSINFO:
+            this->parse_prpsinfo_note(note, attrs);
+            break;
+        default: {
+            colorstream cs;
+            cs << ARION_LOG_COLOR::ORANGE << "Coredump note " << ARION_LOG_COLOR::MAGENTA
+               << int_to_hex(static_cast<int>(note.type())) << ARION_LOG_COLOR::ORANGE << " is not implemented.";
+            arion->logger->warn(cs.str());
+            break;
+        }
+        }
+    }
+    return std::move(elf);
+}
 
 void ElfParser::process()
 {
@@ -30,7 +124,7 @@ void ElfParser::process()
     elf = this->parse_general_data(std::move(elf));
     elf = this->parse_segments(std::move(elf));
     if (elf_attrs->type == ELF_FILE_TYPE::CORE)
-        elf = this->parse_coredump_data(std::move(elf));
+        elf = this->coredump_parser->parse_coredump_data(std::move(elf), elf_attrs, this->segments);
 }
 
 std::unique_ptr<LIEF::ELF::Binary> ElfParser::parse_general_data(std::unique_ptr<LIEF::ELF::Binary> elf)
@@ -68,6 +162,7 @@ std::unique_ptr<LIEF::ELF::Binary> ElfParser::parse_general_data(std::unique_ptr
     default:
         throw UnsupportedCpuArchException();
     }
+
     return std::move(elf);
 }
 
@@ -89,26 +184,6 @@ std::unique_ptr<LIEF::ELF::Binary> ElfParser::parse_segments(std::unique_ptr<LIE
         seg->flags = (lief_seg.has(LIEF::ELF::Segment::FLAGS::R) << 2) |
                      (lief_seg.has(LIEF::ELF::Segment::FLAGS::W) << 1) | lief_seg.has(LIEF::ELF::Segment::FLAGS::X);
         this->segments.push_back(seg);
-    }
-    return std::move(elf);
-}
-
-std::unique_ptr<LIEF::ELF::Binary> ElfParser::parse_coredump_data(std::unique_ptr<LIEF::ELF::Binary> elf)
-{
-    for (const LIEF::ELF::Note &note : elf->notes())
-    {
-        if (LIEF::ELF::CoreFile::classof(&note))
-        {
-            const auto &nt_core_file = static_cast<const LIEF::ELF::CoreFile &>(note);
-            for (auto &file_entry : nt_core_file)
-            {
-                for (std::shared_ptr<struct arion::SEGMENT> &seg : this->segments)
-                {
-                    if (seg->virt_addr >= file_entry.start && seg->virt_addr < file_entry.end)
-                        seg->info = file_entry.path;
-                }
-            }
-        }
     }
     return std::move(elf);
 }